Use BigInteger over BigDecimal

+++ reports-reloaded/src/main/java/com/novoda/github/reports/data/mock/MockRepoDataLayer.java
@@ -3,7 +3,7 @@ package com.novoda.github.reports.data.mock;
 import com.novoda.github.reports.data.RepoDataLayer;
 import com.novoda.github.reports.data.model.ProjectRepoStats;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.Date;
 
 public class MockRepoDataLayer implements RepoDataLayer {
@@ -11,12 +11,12 @@ public class MockRepoDataLayer implements RepoDataLayer {
     public ProjectRepoStats getStats(String repo, Date from, Date to) {
         return new ProjectRepoStats(
                 repo,
-                BigDecimal.valueOf(5),
-                BigDecimal.valueOf(4),
-                BigDecimal.valueOf(5),
-                BigDecimal.valueOf(8),
-                BigDecimal.valueOf(23),
-                BigDecimal.valueOf(4)
+                BigInteger.valueOf(5),
+                BigInteger.valueOf(4),
+                BigInteger.valueOf(5),
+                BigInteger.valueOf(8),
+                BigInteger.valueOf(23),
+                BigInteger.valueOf(4)
         );
     }
 }
